+++ packages/common/src/identity.ts
@@ -25,7 +25,7 @@ export type IpArData = {
     proofComEncEq: string;
 };
 
-export interface IdObjectRequest {
+export interface IdObjectRequestV1 {
     idCredPub: string;
     choiceArData: {
         arIdentities: number[];
@@ -39,14 +39,14 @@ export interface IdObjectRequest {
 }
 
 /**
- * Creates an V1 identityRequest.
+ * Creates a V1 identityRequest.
  */
 export function createIdentityRequest(
     input: IdentityRequestInput
-): IdObjectRequest {
+): Versioned<IdObjectRequestV1> {
     const rawRequest = wasm.createIdRequestV1(JSON.stringify(input));
     try {
-        return JSON.parse(rawRequest).idObjectRequest.value;
+        return JSON.parse(rawRequest).idObjectRequest;
     } catch (e) {
         throw new Error(rawRequest);
     }