+++ packages/nodejs/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-### Added
+### Changed
 
-- `createIdentityRequest`, to create identity requests.
+- Bumped @concordium/common-sdk to 2.4.0.
 
 ## 3.0.2 2022-7-26
 
@@ -41,7 +41,7 @@
 - Support deserializing new schema types: ULeb128, ILeb128, ByteArray and ByteList.
 - Support deserializing schemas with versioning information.
 
-### Changes
+### Changed
 
 - The function for deserializing a module schema `deserialModuleFromBuffer` now have the schema version as an optional argument. The function will try to extract the version from the buffer. When a version is provided it falls back to this, otherwise it throws an error.